+++ crates/wdl-grammar/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Fixed the experimental parser to skip parsing if it cannot find a supported
+  version statement ([#59](https://github.com/stjude-rust-labs/wdl/pull/59))
 * Fixed handling of `None` literal values in expressions in the experimental
   parser ([#58](https://github.com/stjude-rust-labs/wdl/pull/58)).
 * Fixed the experimental parser to accept multiple placeholder options